Cardano.Logging

Documentation

class Contravariant (f :: Type -> Type) where

The class of contravariant functors.

Whereas in Haskell, one can think of a Functor as containing or producing values, a contravariant functor is a functor that can be thought of as consuming values.

As an example, consider the type of predicate functions a -> Bool. One such predicate might be negative x = x < 0, which classifies integers as to whether they are negative. However, given this predicate, we can re-use it in other situations, providing we have a way to map values to integers. For instance, we can use the negative predicate on a person's bank balance to work out if they are currently overdrawn:

newtype Predicate a = Predicate { getPredicate :: a -> Bool }

instance Contravariant Predicate where
  contramap :: (a' -> a) -> (Predicate a -> Predicate a')
  contramap f (Predicate p) = Predicate (p . f)
                                         |   `- First, map the input...
                                         `----- then apply the predicate.

overdrawn :: Predicate Person
overdrawn = contramap personBankBalance negative

Any instance should be subject to the following laws:

Identity

contramap id = id

Composition

contramap (g . f) = contramap f . contramap g

Note, that the second law follows from the free theorem of the type of contramap and the first law, so you need only check that the former condition holds.

Minimal complete definition

contramap

Methods

contramap :: (a' -> a) -> f a -> f a'

(>$) :: b -> f b -> f a infixl 4

Replace all locations in the output with the same value. The default definition is contramap . const, but this may be overridden with a more efficient version.

newtype Trace m a

The Trace carries the underlying tracer Tracer from the contra-tracer package. It adds a LoggingContext and maybe a TraceControl to every message.

Constructors

Trace

Fields

unpackTrace :: Tracer m (LoggingContext, Either TraceControl a)

data TraceDispatcherMessage

Constructors

StartLimiting Text
This message indicates the start of frequency limiting

StopLimiting Text Int
This message indicates the stop of frequency limiting, and gives the number of messages that has been suppressed

RememberLimiting Text Int
This message remembers of ongoing frequency limiting, and gives the number of messages that has been suppressed

UnknownNamespace [Text] [Text] UnknownNamespaceKind
An internal error was detected

TracerInfo [Text] [Text] [Text]
The first array signifies the namespace of silent tracers The second array signifies the namespace tracers without metrics The third array gives the names of all tracers

MetricsInfo (Map Text Int)
Outputs optional statistics about metrics frequency

TracerConsistencyWarnings [Text]
Consistency check found warnings

TracerInfoConfig TraceConfig
Trace the effective configuration as JSON

data DataPoint where

Type wrapper for some value of type v. The only reason we need this wrapper is an ability to store different values in the same DataPointStore.

Please note that when the acceptor application will read the value of type v from the store, this value is just as unstructured JSON, but not Haskell value of type v. That's why FromJSON instance for type v should be available for the acceptor application, to decode unstructured JSON.

Constructors

DataPoint :: (ToJSON v, NFData v) => !v -> DataPoint

data ConfigReflection

Constructors

ConfigReflection

Fields

crSilent :: IORef (Set [Text])

crNoMetrics :: IORef (Set [Text])

crAllTracers :: IORef (Set [Text])

type NSWarnings = [Text]

Warnings as a list of text

data DocuResult

data DocTracer

Constructors

DocTracer

Fields

dtTracerNames :: [[Text]]

dtSilent :: [[Text]]

dtNoMetrics :: [[Text]]

dtBuilderList :: [([Text], DocuResult)]

dtWarnings :: [InconsistencyWarning]

data LimiterSpec

Constructors

LimiterSpec

Fields

lsNs :: [Text]

lsName :: Text

lsFrequency :: Double

data UnknownNamespaceKind

Constructors

UKFSeverity

UKFPrivacy

UKFDetails

type DataPointName = Text

type DataPointStore = TVar (Map DataPointName DataPoint)

class LogFormatting a where

Every message needs this to define how to represent itself

Minimal complete definition

forMachine

Methods

forMachine :: DetailLevel -> a -> Object

Machine readable representation with the possibility to represent with varying serialisations based on the detail level. This will result in JSON formatted log output. A forMachine implementation is required for any instance definition.

forHuman :: a -> Text

Human-readable representation. The empty text indicates there's no specific human-readable formatting for that type - this is the default implementation. If however human-readble output is explicitly requested, e.g. by logs, the system will fall back to a JSON object conforming to the forMachine definition, and rendering it as a value in `{"data": value}`. Leaving out forHuman in some instance definition will not lead to loss of log information that way.

asMetrics :: a -> [Metric]

Metrics representation. The default indicates that no metric is based on trace occurrences of that type.

data Metric

Constructors

IntM Text Integer
An integer metric. Text is used to name the metric

DoubleM Text Double
A double metric. Text is used to name the metric

CounterM Text (Maybe Int)
A counter metric. Text is used to name the metric

PrometheusM Text [(Text, Text)]
A prometheus metric with key label pairs. Text is used to name the metric [(Text, Text)] is used to represent the key label pairs The value of the metric will always be "1" e.g. if you have a prometheus metric with the name "prometheus_metric" and the key label pairs [("key1", "value1"), ("key2", "value2")] the metric will be represented as "prometheus_metric{key1="value1",key2="value2"} 1"

newtype Documented a

Constructors

Documented

Fields

undoc :: [DocMsg a]

data DocMsg a

Document a message by giving a prototype, its most special name in the namespace and a comment in markdown format

Constructors

DocMsg

Fields

dmNamespace :: Namespace a

dmMetricsMD :: [(Text, Text)]

dmMarkdown :: Text

data LoggingContext

Context any log message carries

Constructors

LoggingContext

Fields

lcNSInner :: [Text]

lcNSPrefix :: [Text]

lcSeverity :: Maybe SeverityS

lcPrivacy :: Maybe Privacy

lcDetails :: Maybe DetailLevel

data Namespace a

A unique identifier for every message, composed of text A namespace can as well appear with the tracer name (e.g. ChainDB.OpenEvent.OpenedDB), or more prefixes, in this moment it is a NamespaceOuter is used

Constructors

Namespace

Fields

nsPrefix :: [Text]

nsInner :: [Text]

class MetaTrace a where

Minimal complete definition

namespaceFor, severityFor, documentFor, allNamespaces

Methods

namespaceFor :: a -> Namespace a

severityFor :: Namespace a -> Maybe a -> Maybe SeverityS

privacyFor :: Namespace a -> Maybe a -> Maybe Privacy

detailsFor :: Namespace a -> Maybe a -> Maybe DetailLevel

documentFor :: Namespace a -> Maybe Text

metricsDocFor :: Namespace a -> [(Text, Text)]

allNamespaces :: [Namespace a]

data DetailLevel

Formerly known as verbosity

Constructors

DMinimal

DNormal

DDetailed

DMaximum

data Privacy

Privacy of a message. Default is Public

Constructors

Confidential
confidential information - handle with care

Public
can be public.

data SeverityS

Severity of a message

Constructors

Debug
Debug messages

Info
Information

Notice
Normal runtime Conditions

Warning
General Warnings

Error
General Errors

Critical
Severe situations

Alert
Take immediate action

Emergency
System is unusable

newtype SeverityF

Severity for a filter Nothing means don't show anything (Silence) Nothing level means show messages with severity >= level

Constructors

SeverityF (Maybe SeverityS)

data ConfigOption

Constructors

ConfSeverity
Severity level for a filter (default is Warning)

Fields

severity :: SeverityF

ConfDetail
Detail level (default is DNormal)

Fields

detail :: DetailLevel

ConfBackend
To which backend to pass Default is [EKGBackend, Forwarder, Stdout MachineFormat]

Fields

backends :: [BackendConfig]

ConfLimiter
Construct a limiter with limiting to the Double, which represents frequency in number of messages per second

Fields

maxFrequency :: Double

newtype ForwarderAddr

Constructors

LocalSocket FilePath

data FormatLogging

Constructors

HumanFormatColoured

HumanFormatUncoloured

MachineFormat

data ForwarderMode

Constructors

Initiator
Forwarder works as a client: it initiates network connection with 'cardano-tracer' and/or another Haskell acceptor application.

Responder
Forwarder works as a server: it accepts network connection from 'cardano-tracer' and/or another Haskell acceptor application.

data Verbosity

Constructors

Maximum
Maximum verbosity for all tracers in the forwarding protocols.

Minimum
Minimum verbosity, the forwarding will work as silently as possible.

data TraceOptionForwarder

Constructors

TraceOptionForwarder

Fields

tofQueueSize :: Word

tofVerbosity :: Verbosity

tofMaxReconnectDelay :: Word

data TraceConfig

Constructors

TraceConfig

Fields

tcOptions :: Map [Text] [ConfigOption]

Options specific to a certain namespace

tcForwarder :: Maybe TraceOptionForwarder

Options for the forwarder

tcNodeName :: Maybe Text

Optional human-readable name of the node.

tcMetricsPrefix :: Maybe Text

Optional prefix for metrics.

tcResourceFrequency :: Maybe Int

Optional resource trace frequency in milliseconds.

tcLedgerMetricsFrequency :: Maybe Int

Optional ledger metrics frequency in milliseconds.

data FormattedMessage

Constructors

FormattedHuman Bool Text
The bool specifies if the formatting includes colours

FormattedMachine Text

FormattedMetrics [Metric]

FormattedForwarder TraceObject

FormattedCBOR ByteString

data TraceControl where

When configuring a net of tracers, it should be run with Config on all entry points first, and then with TCOptimize. When reconfiguring it needs to run TCReset followed by Config followed by TCOptimize

Constructors

TCReset :: TraceControl

TCConfig :: TraceConfig -> TraceControl

TCOptimize :: ConfigReflection -> TraceControl

TCDocument :: Int -> DocCollector -> TraceControl

newtype DocCollector

Constructors

DocCollector (IORef (Map Int LogDoc))

data LogDoc

Constructors

LogDoc

Fields

ldDoc :: !Text

ldMetricsDoc :: !(Map Text Text)

ldNamespace :: ![([Text], [Text])]

ldSeverityCoded :: !(Maybe SeverityS)

ldPrivacyCoded :: !(Maybe Privacy)

ldDetailsCoded :: !(Maybe DetailLevel)

ldDetails :: ![DetailLevel]

ldBackends :: ![BackendConfig]

ldFiltered :: ![SeverityF]

ldLimiter :: ![(Text, Double)]

ldSilent :: Bool

data BackendConfig

Constructors

Forwarder

Stdout FormatLogging

EKGBackend

DatapointBackend

PrometheusSimple Bool (Maybe HostName) PortNumber

newtype Folding a b

Type for the function foldTraceM from module CardanoLoggingTrace

Constructors

Folding b

data TraceObject

Used as interface object for ForwarderTracer

Constructors

TraceObject

Fields

toHuman :: !(Maybe Text)

toMachine :: !Text

toNamespace :: ![Text]

toSeverity :: !SeverityS

toDetails :: !DetailLevel

toTimestamp :: !UTCTime

toHostname :: !Text

toThreadId :: !Text

data PreFormatted

Constructors

PreFormatted

Fields

pfTime :: !UTCTime

pfNamespace :: !Text

pfThreadId :: !Text

pfForHuman :: !(Maybe Text)

pfForMachineObject :: Object

data HowToConnect

Constructors

LocalPipe !FilePath
Local pipe (UNIX or Windows).

RemoteSocket !Host !Port
Remote socket (host and port).

traceWith :: Monad m => Trace m a -> a -> m ()

Adds a message object to a trace

unfold :: Folding a b -> b

runTracer :: Tracer m a -> TracerA m a ()

contramapM :: Monad m => Trace m b -> ((LoggingContext, Either TraceControl a) -> m (LoggingContext, Either TraceControl b)) -> m (Trace m a)

Contramap a monadic function over a trace

natTracer :: forall m n s. (forall x. m x -> n x) -> Tracer m s -> Tracer n s

Use a natural transformation to change the m type. This is useful, for instance, to use concrete IO tracers in monad transformer stacks that have IO as their base.

stdoutTracer :: Tracer IO String

Trace strings to stdout. Output could be jumbled when this is used from multiple threads. Consider debugTracer instead.

debugTracer :: Applicative m => Tracer m String

Trace strings using traceM. This will use stderr. See documentation in Debug.Trace for more details.

arrow :: TracerA m a () -> Tracer m a

Inverse of use.

emptyConfigReflection :: IO ConfigReflection

configureTracers :: forall a m. (MetaTrace a, MonadIO m) => ConfigReflection -> TraceConfig -> [Trace m a] -> m ()

Call this function at initialisation, and later for reconfiguration. Config reflection is used to optimise the tracers and has to collect information about the tracers. Although it is possible to give more then one tracer of the same time, it is not a common case to do this.

withNamespaceConfig :: forall m a b c. (MonadIO m, Ord b) => String -> (TraceConfig -> Namespace a -> m b) -> (Maybe b -> Trace m c -> m (Trace m a)) -> Trace m c -> m (Trace m a)

Take a selector function called extract. Take a function from trace to trace with this config dependent value. In this way construct a trace transformer with a config value

filterSeverityFromConfig :: MonadIO m => Trace m a -> m (Trace m a)

Filter a trace by severity and take the filter value from the config

withDetailsFromConfig :: MonadIO m => Trace m a -> m (Trace m a)

Set detail level of a trace from the config

withBackendsFromConfig :: MonadIO m => (Maybe [BackendConfig] -> Trace m FormattedMessage -> m (Trace m a)) -> m (Trace m a)

Routing and formatting of a trace from the config

withLimitersFromConfig :: forall a m. MonadUnliftIO m => Trace m TraceDispatcherMessage -> Trace m a -> m (Trace m a)

Routing and formatting of a trace from the config

maybeSilent :: forall m a. MonadIO m => (TraceConfig -> Namespace a -> Bool) -> [Text] -> Bool -> Trace m a -> m (Trace m a)

Switch off any message of a particular tracer based on the configuration. If the top tracer is silent and no subtracer is not silent, then switch it off

isSilentTracer :: forall a. MetaTrace a => TraceConfig -> Namespace a -> Bool

hasNoMetrics :: forall a. MetaTrace a => TraceConfig -> Namespace a -> Bool

getSeverity :: TraceConfig -> Namespace a -> SeverityF

If no severity can be found in the config, it is set to Warning

getDetails :: TraceConfig -> Namespace a -> DetailLevel

If no details can be found in the config, it is set to DNormal

getBackends :: TraceConfig -> Namespace a -> [BackendConfig]

If no backends can be found in the config, it is set to [EKGBackend, Forwarder, Stdout HumanFormatColoured]

addFiltered :: MonadIO m => TraceControl -> Maybe SeverityF -> m ()

Callback for doc collection

addLimiter :: MonadIO m => TraceControl -> (Text, Double) -> m ()

Callback for doc collection

addSilent :: MonadIO m => TraceControl -> Maybe Bool -> m ()

limitFrequency :: forall a m. MonadUnliftIO m => Double -> Text -> Trace m TraceDispatcherMessage -> Trace m a -> m (Trace m a)

Limits the frequency of messages to nMsg which is given per minute.

If the limiter detects more messages, it traces randomly selected messages with the given frequency on the vtracer until the frequency falls under the threshold long enough.(see below)

Before this the ltracer gets a StartLimiting message. In-between you receive ContinueLimiting messages on the ltracer every reminderPeriod seconds, with the number of suppressed messages. Finally it sends a StopLimiting message on the ltracer and traces all messages on the vtracer again.

A budget is used to decide when to start limiting and stop limiting, so that the limiter does not get activated if few messages are send in high frequency, and doesn't get deactivated if their are only few messages which come with low frequency. When messages arrive in shorter frequency then by the given thresholdFrequency budget is earned, and if they arrive in a longer period budget is spend. If budget is gets higher then budgetLimit, the limiter starts, and if it falls below minus budgetLimit the limiter stops.

readConfiguration :: FilePath -> IO TraceConfig

Read a configuration file and returns the internal representation

readConfigurationWithDefault :: FilePath -> TraceConfig -> IO TraceConfig

Read a configuration file and returns the internal representation Uses values which are not in the file from the defaultConfig

configToRepresentation :: TraceConfig -> ConfigRepresentation

Convert from internal to external representation

checkTraceConfiguration :: FilePath -> TraceConfig -> [([Text], [Text])] -> IO NSWarnings

Checks if all namespaces in this configuration are legal. Legal in this case means that it can be found by a hierarchcical lookup in all namespaces. Warns if namespaces in all namespaces are not unique, Warns if namespaces in all namespaces are ending in the middle of another namespace. The namespaces in allNamespaces are consistent with the namespaces for the severityFor, privacyFor, detailsFor, documentFor and metricsDofFor functions.

checkTraceConfiguration' :: TraceConfig -> [([Text], [Text])] -> NSWarnings

documentTracer :: forall a. MetaTrace a => Trace IO a -> IO DocTracer

documentTracer' :: forall a a1. MetaTrace a => (Trace IO a1 -> IO (Trace IO a)) -> Trace IO a1 -> IO DocTracer

docuResultsToText :: DocTracer -> TraceConfig -> Text

docuResultsToMetricsHelptext :: DocTracer -> Text

docTracer :: MonadIO m => BackendConfig -> Trace m FormattedMessage

docTracerDatapoint :: MonadIO m => BackendConfig -> Trace m a

docIt :: MonadIO m => BackendConfig -> (LoggingContext, Either TraceControl a) -> m ()

Callback for doc collection

addDocumentedNamespace :: [Text] -> Documented a -> Documented a

Convenience function for adding a namespace prefix to a documented

metricsFormatter :: forall a m. (LogFormatting a, MonadIO m) => Trace m FormattedMessage -> Trace m a

Format this trace as metrics

preFormatted :: (LogFormatting a, MonadIO m) => Bool -> Trace m PreFormatted -> m (Trace m a)

Transform this trace to a preformatted message, so that double serialization is avoided

forwardFormatter :: forall a m. MonadIO m => LogFormatting a => Trace m FormattedMessage -> m (Trace m a)

forwardFormatter' :: forall m. MonadIO m => Trace m FormattedMessage -> Trace m PreFormatted

Format this trace as TraceObject for the trace forwarder

machineFormatter :: forall a m. (MonadIO m, LogFormatting a) => Trace m FormattedMessage -> m (Trace m a)

machineFormatter' :: forall m. MonadIO m => Trace m FormattedMessage -> Trace m PreFormatted

Format this trace as TraceObject for machine-readable text output (JSON)

cborFormatter :: forall a m. (MonadIO m, LogFormatting a) => Trace m FormattedMessage -> m (Trace m a)

cborFormatter' :: forall m. MonadIO m => Trace m FormattedMessage -> Trace m PreFormatted

Format this trace in binary serialisation (CBOR)

humanFormatter :: forall a m. MonadIO m => LogFormatting a => Bool -> Trace m FormattedMessage -> m (Trace m a)

humanFormatter' :: forall m. MonadIO m => Bool -> Trace m FormattedMessage -> Trace m PreFormatted

Format this trace in human readable text output

threadLabelMe :: String -> IO ()

runInLoop :: IO () -> (SomeException -> IO ()) -> Word -> Word -> IO ()

Run an IO action which may throw an exception in a loop. On exception, the action will be re-run after a pause. That pause doubles which each exception, but is reset when the action runs long enough.

showT :: Show a => a -> Text

Convenience function for a Show instance to be converted to text immediately

withLoggingContext :: Monad m => LoggingContext -> Trace m a -> Trace m a

Sets a new logging context for this message

filterTrace :: Monad m => ((LoggingContext, a) -> Bool) -> Trace m a -> Trace m a

filterTraceMaybe :: Monad m => Trace m a -> Trace m (Maybe a)

filterTraceBySeverity :: Monad m => Maybe SeverityF -> Trace m a -> Trace m a

filterTraceByPrivacy :: Monad m => Maybe Privacy -> Trace m a -> Trace m a

setSeverity :: Monad m => SeverityS -> Trace m a -> Trace m a

Sets severity for the messages in this trace

withSeverity :: forall m a. (Monad m, MetaTrace a) => Trace m a -> Trace m a

Sets severities for the messages in this trace based on the MetaTrace class

privately :: Monad m => Trace m a -> Trace m a

Sets privacy Confidential for the messages in this trace

setPrivacy :: Monad m => Privacy -> Trace m a -> Trace m a

Sets privacy for the messages in this trace

withPrivacy :: forall m a. (Monad m, MetaTrace a) => Trace m a -> Trace m a

Sets privacy for the messages in this trace based on the MetaTrace class

allPublic :: a -> Privacy

allConfidential :: a -> Privacy

setDetails :: Monad m => DetailLevel -> Trace m a -> Trace m a

Sets detail level for the messages in this trace

withDetails :: forall m a. (Monad m, MetaTrace a) => Trace m a -> Trace m a

Sets detail level for the messages in this trace based on the message

contramapMCond :: Monad m => Trace m b -> ((LoggingContext, Either TraceControl a) -> m (Maybe (LoggingContext, Either TraceControl b))) -> m (Trace m a)

Contramap a monadic function over a trace Can as well filter out messages

contramapM' :: Monad m => ((LoggingContext, Either TraceControl a) -> m ()) -> Trace m a

foldTraceM :: forall a acc m. MonadUnliftIO m => (acc -> LoggingContext -> a -> m acc) -> acc -> Trace m (Folding a acc) -> m (Trace m a)

Folds the monadic cata function with acc over a. Uses an MVar to store the state

foldCondTraceM :: forall a acc m. MonadUnliftIO m => (acc -> LoggingContext -> a -> m acc) -> acc -> (a -> Bool) -> Trace m (Folding a acc) -> m (Trace m a)

Like foldTraceM, but filter the trace by a predicate.

routingTrace :: forall m a. Monad m => (a -> m (Trace m a)) -> Trace m a -> Trace m a

Allows to route to different tracers, based on the message being processed. The second argument must mappend all possible tracers of the first argument to one tracer. This is required for the configuration!

withNames :: forall m a. (Monad m, MetaTrace a) => [Text] -> Trace m a -> Trace m a

Sets names for the messages in this trace based on the selector function and appends the provided names to the context.

appendPrefixName :: Monad m => Text -> Trace m a -> Trace m a

Appends a name to the context. E.g. appendName "specific" $ appendName "middle" $ appendName "general" tracer give the result: `general.middle.specific`.

appendPrefixNames :: Monad m => [Text] -> Trace m a -> Trace m a

Appends all names to the context.

appendInnerName :: Monad m => Text -> Trace m a -> Trace m a

appendInnerNames :: Monad m => [Text] -> Trace m a -> Trace m a

Appends all names to the context.

withInnerNames :: forall m a. (Monad m, MetaTrace a) => Trace m a -> Trace m a

Sets names for the messages in this trace based on the selector function

contramap' :: Contravariant f => (a' -> a) -> f a -> f a'

A contramap' which is strict in its second argument and its result captures a common pattern to avoid unintentionally leaking space when composing tracers. The infix alias is (>!$!<).

(>!$!<) :: Contravariant f => (a' -> a) -> f a -> f a' infixl 4

A contramap' which is strict in its second argument and its result captures a common pattern to avoid unintentionally leaking space when composing tracers. The infix alias is (>!$!<).

initDataPointStore :: IO DataPointStore

writeToStore :: DataPointStore -> DataPointName -> DataPoint -> IO ()

Write DataPoint to the store.

dataPointTracer :: forall m. MonadIO m => DataPointStore -> Trace m DataPoint

mkDataPointTracer :: forall dp. (ToJSON dp, MetaTrace dp, NFData dp) => Trace IO DataPoint -> IO (Trace IO dp)

ekgTracer :: MonadIO m => TraceConfig -> Store -> m (Trace m FormattedMessage)

It is mandatory to construct only one standard tracer in any application! Throwing away a standard tracer and using a new one will result in an exception

showTReal :: RealFloat a => a -> Text

standardTracer :: forall m. MonadIO m => m (Trace m FormattedMessage)

The standardTracer handles stdout logging in a thread-safe manner. It is strongly advised to construct only one standardTracer for any application.

forwardTracer :: forall m. MonadIO m => (TraceObject -> IO ()) -> Trace m FormattedMessage

It is mandatory to construct only one forwardTracer tracer in any application! Throwing away a forwardTracer tracer and using a new one will result in an exception

mkCardanoTracer :: forall evt. (LogFormatting evt, MetaTrace evt) => Trace IO FormattedMessage -> Trace IO FormattedMessage -> Maybe (Trace IO FormattedMessage) -> [Text] -> IO (Trace IO evt)

Construct a tracer according to the requirements for cardano node. The tracer gets a name, which is appended to its namespace. The tracer has to be an instance of LogFormatting for the display of messages and an instance of MetaTrace for meta information such as severity, privacy, details and backends'. The tracer gets the backends': trStdout, trForward and mbTrEkg as arguments. The returned tracer needs to be configured with a configuration before it is used.

mkCardanoTracer' :: forall evt evt1. (LogFormatting evt1, MetaTrace evt1) => Trace IO FormattedMessage -> Trace IO FormattedMessage -> Maybe (Trace IO FormattedMessage) -> [Text] -> (Trace IO evt1 -> IO (Trace IO evt)) -> IO (Trace IO evt)

Adds the possibility to add special tracers via the hook function

mkMetricsTracer :: Maybe (Trace IO FormattedMessage) -> Trace IO FormattedMessage

traceTracerInfo :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> ConfigReflection -> IO ()

traceConfigWarnings :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> [Text] -> IO ()

traceEffectiveConfiguration :: Trace IO FormattedMessage -> Trace IO FormattedMessage -> TraceConfig -> IO ()

getMetricName :: Metric -> Text

emptyObject :: HashMap Text a

A helper function for creating an empty |Object|.

emptyLoggingContext :: LoggingContext

nsReplacePrefix :: [Text] -> Namespace a -> Namespace a

nsReplaceInner :: [Text] -> Namespace a -> Namespace a

nsCast :: Namespace a -> Namespace b

nsPrependInner :: Text -> Namespace a -> Namespace b

nsGetComplete :: Namespace a -> [Text]

nsGetTuple :: Namespace a -> ([Text], [Text])

nsRawToText :: ([Text], [Text]) -> Text

nsToText :: Namespace a -> Text

defaultForwarder :: TraceOptionForwarder

emptyTraceConfig :: TraceConfig

emptyLogDoc :: Text -> [(Text, Text)] -> LogDoc

emit :: Applicative m => (a -> m ()) -> TracerA m a ()

Do an emitting effect. Contrast with effect which does not make the tracer an emitting tracer.

effect :: (a -> m b) -> TracerA m a b

Do a non-emitting effect. This effect will only be run if some part of the tracer downstream emits (see emit).

squelch :: Applicative m => TracerA m a ()

Ignore the input and do not emit. The name is intended to lead to clear and suggestive arrow expressions.

nat :: (forall x. m x -> n x) -> TracerA m a b -> TracerA n a b

Use a natural transformation to change the underlying monad.

use :: Tracer m a -> TracerA m a ()

Inverse of arrow. Useful when writing arrow tracers which use a contravariant tracer (the newtype in this module).

traceMaybe :: Monad m => (a -> Maybe b) -> Tracer m b -> Tracer m a

Run a tracer only for the Just variant of a Maybe. If it's Nothing, the nullTracer is used (no output).

The arrow representation allows for proper laziness: if the tracer parameter does not produce any tracing effects, then the predicate won't even be evaluated. Contrast with the simple contravariant representation as a -> m (), in which the predicate _must_ be forced no matter what, because it's impossible to know a priori whether that function will not produce any tracing effects.

It's written out explicitly for demonstration. Could also use arrow notation:

traceMaybe p tr = Tracer $ proc a -> do
  case k a of
    Just b  -> use tr        -< b
    Nothing -> Arrow.squelch -< ()

squelchUnless :: Monad m => (a -> Bool) -> Tracer m a -> Tracer m a

Uses traceMaybe to give a tracer which emits only if a predicate is true.

showTHex :: Integral a => a -> Text