Cardano.Logging.ConfigurationParser

Documentation

readConfiguration :: FilePath -> IO TraceConfig

Read a configuration file and returns the internal representation

readConfigurationWithDefault :: FilePath -> TraceConfig -> IO TraceConfig

Read a configuration file and returns the internal representation Uses values which are not in the file from the defaultConfig

configToRepresentation :: TraceConfig -> ConfigRepresentation

Convert from internal to external representation

Orphan instances

ToJSON TraceConfig

Instance details

Methods

toJSON :: TraceConfig -> Value

toEncoding :: TraceConfig -> Encoding

toJSONList :: [TraceConfig] -> Value

toEncodingList :: [TraceConfig] -> Encoding

omitField :: TraceConfig -> Bool