Cardano.Logging.DocuGenerator.Tree

Documentation

data Tree a

Non-empty, possibly infinite, multi-way trees; also known as rose trees.

Constructors

Node

Fields

rootLabel :: a

label value

subForest :: [Tree a]

zero or more child trees

foldTree :: (a -> [b] -> b) -> Tree a -> b

Fold a tree into a "summary" value in depth-first order.

For each node in the tree, apply f to the rootLabel and the result of applying f to each subForest.

This is also known as the catamorphism on trees.

Examples

Expand

Sum the values in a tree:

foldTree (\x xs -> sum (x:xs)) (Node 1 [Node 2 [], Node 3 []]) == 6

Find the maximum value in the tree:

foldTree (\x xs -> maximum (x:xs)) (Node 1 [Node 2 [], Node 3 []]) == 3

Count the number of leaves in the tree:

foldTree (\_ xs -> if null xs then 1 else sum xs) (Node 1 [Node 2 [], Node 3 []]) == 2

Find depth of the tree; i.e. the number of branches from the root of the tree to the furthest leaf:

foldTree (\_ xs -> if null xs then 0 else 1 + maximum xs) (Node 1 [Node 2 [], Node 3 []]) == 1

You can even implement traverse using foldTree:

traverse' f = foldTree (\x xs -> liftA2 Node (f x) (sequenceA xs))

Since: containers-0.5.8

printTree :: Tree Text -> Text

printList :: (a -> Builder) -> [a] -> Builder

toForest :: [[Text]] -> Forest Text