Cardano.Logging.Prometheus.NetworkRun

Description

Run a TCP server, with hardening against connection flooding

Documentation

data NetworkRunParams

Parameters specifying how the TCP server should be run

Constructors

NetworkRunParams

Fields

runSocketTimeout :: !Int

Release socket after inactivity (seconds)

runSocketGraceful :: !Int

Graceful closing of socket (milliseconds), 0 to disable

runRecvMaxSize :: !Int

Close socket if more than (runRecvMaxSize - 1) bytes received; choose a small power of 2

runRateLimit :: !Double

Limit requests per second (may be < 0.0), 0.0 to disable

runConnLimitGlobal :: !Int

Limit total number of incoming connections, 0 to disable

runConnLimitPerHost :: !Int

Limit number of incoming connections from the same host, 0 to disable

runServerName :: !String

The server name - exclusively used for labeling GHC threads

type TimeoutServer a = NetworkRunParams -> IO () -> Socket -> IO a

defaultRunParams :: String -> NetworkRunParams

mkTCPServerRunner :: NetworkRunParams -> Maybe HostName -> PortNumber -> TimeoutServer () -> IO (IO ())

Returns an IO action that will run a TCP server conforming to the run parameters. Will bind to localhost / loopback device only if no host name is specified. Will throw an exception when TCP server startup fails - the caller is responsible for appropriately reacting to that.