Cardano.Logging.Trace

Documentation

traceWith :: Monad m => Trace m a -> a -> m ()

Adds a message object to a trace

withLoggingContext :: Monad m => LoggingContext -> Trace m a -> Trace m a

Sets a new logging context for this message

filterTrace :: Monad m => ((LoggingContext, a) -> Bool) -> Trace m a -> Trace m a

filterTraceMaybe :: Monad m => Trace m a -> Trace m (Maybe a)

filterTraceBySeverity :: Monad m => Maybe SeverityF -> Trace m a -> Trace m a

filterTraceByPrivacy :: Monad m => Maybe Privacy -> Trace m a -> Trace m a

setSeverity :: Monad m => SeverityS -> Trace m a -> Trace m a

Sets severity for the messages in this trace

withSeverity :: forall m a. (Monad m, MetaTrace a) => Trace m a -> Trace m a

Sets severities for the messages in this trace based on the MetaTrace class

privately :: Monad m => Trace m a -> Trace m a

Sets privacy Confidential for the messages in this trace

setPrivacy :: Monad m => Privacy -> Trace m a -> Trace m a

Sets privacy for the messages in this trace

withPrivacy :: forall m a. (Monad m, MetaTrace a) => Trace m a -> Trace m a

Sets privacy for the messages in this trace based on the MetaTrace class

allPublic :: a -> Privacy

allConfidential :: a -> Privacy

setDetails :: Monad m => DetailLevel -> Trace m a -> Trace m a

Sets detail level for the messages in this trace

withDetails :: forall m a. (Monad m, MetaTrace a) => Trace m a -> Trace m a

Sets detail level for the messages in this trace based on the message

contramapM :: Monad m => Trace m b -> ((LoggingContext, Either TraceControl a) -> m (LoggingContext, Either TraceControl b)) -> m (Trace m a)

Contramap a monadic function over a trace

contramapMCond :: Monad m => Trace m b -> ((LoggingContext, Either TraceControl a) -> m (Maybe (LoggingContext, Either TraceControl b))) -> m (Trace m a)

Contramap a monadic function over a trace Can as well filter out messages

contramapM' :: Monad m => ((LoggingContext, Either TraceControl a) -> m ()) -> Trace m a

foldTraceM :: forall a acc m. MonadUnliftIO m => (acc -> LoggingContext -> a -> m acc) -> acc -> Trace m (Folding a acc) -> m (Trace m a)

Folds the monadic cata function with acc over a. Uses an MVar to store the state

foldCondTraceM :: forall a acc m. MonadUnliftIO m => (acc -> LoggingContext -> a -> m acc) -> acc -> (a -> Bool) -> Trace m (Folding a acc) -> m (Trace m a)

Like foldTraceM, but filter the trace by a predicate.

routingTrace :: forall m a. Monad m => (a -> m (Trace m a)) -> Trace m a -> Trace m a

Allows to route to different tracers, based on the message being processed. The second argument must mappend all possible tracers of the first argument to one tracer. This is required for the configuration!

withNames :: forall m a. (Monad m, MetaTrace a) => [Text] -> Trace m a -> Trace m a

Sets names for the messages in this trace based on the selector function and appends the provided names to the context.

appendPrefixName :: Monad m => Text -> Trace m a -> Trace m a

Appends a name to the context. E.g. appendName "specific" $ appendName "middle" $ appendName "general" tracer give the result: `general.middle.specific`.

appendPrefixNames :: Monad m => [Text] -> Trace m a -> Trace m a

Appends all names to the context.

appendInnerName :: Monad m => Text -> Trace m a -> Trace m a

appendInnerNames :: Monad m => [Text] -> Trace m a -> Trace m a

Appends all names to the context.

withInnerNames :: forall m a. (Monad m, MetaTrace a) => Trace m a -> Trace m a

Sets names for the messages in this trace based on the selector function

contramap' :: Contravariant f => (a' -> a) -> f a -> f a'

A contramap' which is strict in its second argument and its result captures a common pattern to avoid unintentionally leaking space when composing tracers. The infix alias is (>!$!<).

(>!$!<) :: Contravariant f => (a' -> a) -> f a -> f a' infixl 4

A contramap' which is strict in its second argument and its result captures a common pattern to avoid unintentionally leaking space when composing tracers. The infix alias is (>!$!<).