Cardano.Logging.Types

Documentation

newtype Trace m a

The Trace carries the underlying tracer Tracer from the contra-tracer package. It adds a LoggingContext and maybe a TraceControl to every message.

Constructors

Trace

Fields

unpackTrace :: Tracer m (LoggingContext, Either TraceControl a)

class LogFormatting a where

Every message needs this to define how to represent itself

Minimal complete definition

forMachine

Methods

forMachine :: DetailLevel -> a -> Object

Machine readable representation with the possibility to represent with varying serialisations based on the detail level. This will result in JSON formatted log output. A forMachine implementation is required for any instance definition.

forHuman :: a -> Text

Human-readable representation. The empty text indicates there's no specific human-readable formatting for that type - this is the default implementation. If however human-readble output is explicitly requested, e.g. by logs, the system will fall back to a JSON object conforming to the forMachine definition, and rendering it as a value in `{"data": value}`. Leaving out forHuman in some instance definition will not lead to loss of log information that way.

asMetrics :: a -> [Metric]

Metrics representation. The default indicates that no metric is based on trace occurrences of that type.

data Metric

Constructors

IntM Text Integer
An integer metric. Text is used to name the metric

DoubleM Text Double
A double metric. Text is used to name the metric

CounterM Text (Maybe Int)
A counter metric. Text is used to name the metric

PrometheusM Text [(Text, Text)]
A prometheus metric with key label pairs. Text is used to name the metric [(Text, Text)] is used to represent the key label pairs The value of the metric will always be "1" e.g. if you have a prometheus metric with the name "prometheus_metric" and the key label pairs [("key1", "value1"), ("key2", "value2")] the metric will be represented as "prometheus_metric{key1="value1",key2="value2"} 1"

getMetricName :: Metric -> Text

emptyObject :: HashMap Text a

A helper function for creating an empty |Object|.

newtype Documented a

Constructors

Documented

Fields

undoc :: [DocMsg a]

data DocMsg a

Document a message by giving a prototype, its most special name in the namespace and a comment in markdown format

Constructors

DocMsg

Fields

dmNamespace :: Namespace a

dmMetricsMD :: [(Text, Text)]

dmMarkdown :: Text

data LoggingContext

Context any log message carries

Constructors

LoggingContext

Fields

lcNSInner :: [Text]

lcNSPrefix :: [Text]

lcSeverity :: Maybe SeverityS

lcPrivacy :: Maybe Privacy

lcDetails :: Maybe DetailLevel

emptyLoggingContext :: LoggingContext

data Namespace a

A unique identifier for every message, composed of text A namespace can as well appear with the tracer name (e.g. ChainDB.OpenEvent.OpenedDB), or more prefixes, in this moment it is a NamespaceOuter is used

Constructors

Namespace

Fields

nsPrefix :: [Text]

nsInner :: [Text]

nsReplacePrefix :: [Text] -> Namespace a -> Namespace a

nsReplaceInner :: [Text] -> Namespace a -> Namespace a

nsCast :: Namespace a -> Namespace b

nsPrependInner :: Text -> Namespace a -> Namespace b

nsGetComplete :: Namespace a -> [Text]

nsGetTuple :: Namespace a -> ([Text], [Text])

nsRawToText :: ([Text], [Text]) -> Text

nsToText :: Namespace a -> Text

class MetaTrace a where

Minimal complete definition

namespaceFor, severityFor, documentFor, allNamespaces

Methods

namespaceFor :: a -> Namespace a

severityFor :: Namespace a -> Maybe a -> Maybe SeverityS

privacyFor :: Namespace a -> Maybe a -> Maybe Privacy

detailsFor :: Namespace a -> Maybe a -> Maybe DetailLevel

documentFor :: Namespace a -> Maybe Text

metricsDocFor :: Namespace a -> [(Text, Text)]

allNamespaces :: [Namespace a]

data DetailLevel

Formerly known as verbosity

Constructors

DMinimal

DNormal

DDetailed

DMaximum

data Privacy

Privacy of a message. Default is Public

Constructors

Confidential
confidential information - handle with care

Public
can be public.

data SeverityS

Severity of a message

Constructors

Debug
Debug messages

Info
Information

Notice
Normal runtime Conditions

Warning
General Warnings

Error
General Errors

Critical
Severe situations

Alert
Take immediate action

Emergency
System is unusable

newtype SeverityF

Severity for a filter Nothing means don't show anything (Silence) Nothing level means show messages with severity >= level

Constructors

SeverityF (Maybe SeverityS)

data ConfigOption

Constructors

ConfSeverity
Severity level for a filter (default is Warning)

Fields

severity :: SeverityF

ConfDetail
Detail level (default is DNormal)

Fields

detail :: DetailLevel

ConfBackend
To which backend to pass Default is [EKGBackend, Forwarder, Stdout MachineFormat]

Fields

backends :: [BackendConfig]

ConfLimiter
Construct a limiter with limiting to the Double, which represents frequency in number of messages per second

Fields

maxFrequency :: Double

newtype ForwarderAddr

Constructors

LocalSocket FilePath

data FormatLogging

Constructors

HumanFormatColoured

HumanFormatUncoloured

MachineFormat

data ForwarderMode

Constructors

Initiator
Forwarder works as a client: it initiates network connection with 'cardano-tracer' and/or another Haskell acceptor application.

Responder
Forwarder works as a server: it accepts network connection from 'cardano-tracer' and/or another Haskell acceptor application.

data Verbosity

Constructors

Maximum
Maximum verbosity for all tracers in the forwarding protocols.

Minimum
Minimum verbosity, the forwarding will work as silently as possible.

data TraceOptionForwarder

Constructors

TraceOptionForwarder

Fields

tofQueueSize :: Word

tofVerbosity :: Verbosity

tofMaxReconnectDelay :: Word

defaultForwarder :: TraceOptionForwarder

data ConfigReflection

Constructors

ConfigReflection

Fields

crSilent :: IORef (Set [Text])

crNoMetrics :: IORef (Set [Text])

crAllTracers :: IORef (Set [Text])

emptyConfigReflection :: IO ConfigReflection

data TraceConfig

Constructors

TraceConfig

Fields

tcOptions :: Map [Text] [ConfigOption]

Options specific to a certain namespace

tcForwarder :: Maybe TraceOptionForwarder

Options for the forwarder

tcNodeName :: Maybe Text

Optional human-readable name of the node.

tcMetricsPrefix :: Maybe Text

Optional prefix for metrics.

tcResourceFrequency :: Maybe Int

Optional resource trace frequency in milliseconds.

tcLedgerMetricsFrequency :: Maybe Int

Optional ledger metrics frequency in milliseconds.

emptyTraceConfig :: TraceConfig

data FormattedMessage

Constructors

FormattedHuman Bool Text
The bool specifies if the formatting includes colours

FormattedMachine Text

FormattedMetrics [Metric]

FormattedForwarder TraceObject

FormattedCBOR ByteString

data TraceControl where

When configuring a net of tracers, it should be run with Config on all entry points first, and then with TCOptimize. When reconfiguring it needs to run TCReset followed by Config followed by TCOptimize

Constructors

TCReset :: TraceControl

TCConfig :: TraceConfig -> TraceControl

TCOptimize :: ConfigReflection -> TraceControl

TCDocument :: Int -> DocCollector -> TraceControl

newtype DocCollector

Constructors

DocCollector (IORef (Map Int LogDoc))

data LogDoc

Constructors

LogDoc

Fields

ldDoc :: !Text

ldMetricsDoc :: !(Map Text Text)

ldNamespace :: ![([Text], [Text])]

ldSeverityCoded :: !(Maybe SeverityS)

ldPrivacyCoded :: !(Maybe Privacy)

ldDetailsCoded :: !(Maybe DetailLevel)

ldDetails :: ![DetailLevel]

ldBackends :: ![BackendConfig]

ldFiltered :: ![SeverityF]

ldLimiter :: ![(Text, Double)]

ldSilent :: Bool

emptyLogDoc :: Text -> [(Text, Text)] -> LogDoc

data BackendConfig

Constructors

Forwarder

Stdout FormatLogging

EKGBackend

DatapointBackend

PrometheusSimple Bool (Maybe HostName) PortNumber

newtype Folding a b

Type for the function foldTraceM from module CardanoLoggingTrace

Constructors

Folding b

unfold :: Folding a b -> b

data TraceObject

Used as interface object for ForwarderTracer

Constructors

TraceObject

Fields

toHuman :: !(Maybe Text)

toMachine :: !Text

toNamespace :: ![Text]

toSeverity :: !SeverityS

toDetails :: !DetailLevel

toTimestamp :: !UTCTime

toHostname :: !Text

toThreadId :: !Text

data PreFormatted

Constructors

PreFormatted

Fields

pfTime :: !UTCTime

pfNamespace :: !Text

pfThreadId :: !Text

pfForHuman :: !(Maybe Text)

pfForMachineObject :: Object

data HowToConnect

Constructors

LocalPipe !FilePath
Local pipe (UNIX or Windows).

RemoteSocket !Host !Port
Remote socket (host and port).