Cardano.Logging.Types.NodePeers

Documentation

newtype NodePeers

This type contains an information about current peers of the node. It will be asked by external applications as a DataPoint.

Constructors

NodePeers [PeerInfoPP]

type PeerInfoPP = Text