Cardano.Logging.Utils

Documentation

threadLabelMe :: String -> IO ()

runInLoop :: IO () -> (SomeException -> IO ()) -> Word -> Word -> IO ()

Run an IO action which may throw an exception in a loop. On exception, the action will be re-run after a pause. That pause doubles which each exception, but is reset when the action runs long enough.

showT :: Show a => a -> Text

Convenience function for a Show instance to be converted to text immediately

showTReal :: RealFloat a => a -> Text

showTHex :: Integral a => a -> Text