Cardano.Logging.Utils

Documentation

runInLoop :: IO () -> FilePath -> Word -> IO ()

Run monadic action in a loop. If there's an exception, it will re-run the action again, after pause that grows.

uncurry3 :: (a -> b -> c -> d) -> (a, b, c) -> d

Converts a curried function to a function on a triple.

showT :: Show a => a -> Text

Convenience function for a Show instance to be converted to text immediately

showTHex :: (Integral a, Show a) => a -> Text

Convenience function for a showHex call converted to text immediately